+++ frontend/src/generated-sources/ai/giskard/web/dto/slice-suggestion-dto.ts
@@ -1,0 +1,9 @@
+/**
+ * Generated from ai.giskard.web.dto.SliceSuggestionDTO
+ */
+export interface SliceSuggestionDTO {
+    column: string;
+    operator: string;
+    projectId: number;
+    value: string;
+}
