+++ python-client/giskard/cli_utils.py
@@ -9,10 +9,19 @@ from time import sleep
 from lockfile.pidlockfile import PIDLockFile
 from pydantic import AnyHttpUrl, parse_obj_as
 
+from giskard.client.analytics_collector import GiskardAnalyticsCollector
 from giskard.path_utils import run_dir
 from giskard.settings import settings
 
 logger = logging.getLogger(__name__)
+logging.getLogger().setLevel(logging.INFO)
+logging.getLogger("giskard").setLevel(logging.INFO)
+analytics = GiskardAnalyticsCollector()
+
+
+def set_verbose(_ctx, _param, value):
+    if value:
+        logging.getLogger().setLevel(logging.DEBUG)
 
 
 def remove_stale_pid_file(pid_file):
